requirements.txt
blinker==1.8.1
boto3==1.34.95
botocore==1.34.95
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
werkzeug==3.0.2
zipp==3.18.1